refactor(mini-swe): split v2 sandbox agent
