+++ responses_api_agents/mini_swe_agent_2/.gitignore
@@ -1,0 +1,1 @@
+results/
