+++ responses_api_agents/mini_swe_agent_2/client.py
@@ -1,0 +1,37 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import json
+from asyncio import run
+
+from nemo_gym.server_utils import ServerClient
+
+
+server_client = ServerClient.load_from_global_config()
+task = server_client.post(
+    server_name="mini_swe_simple_agent",
+    url_path="/run",
+    json={
+        "responses_create_params": {
+            "temperature": 0.0,
+            "top_p": 1.0,
+            "input": [],
+        },
+        "instance_id": "pydantic__pydantic-5506",
+        "subset": "gym",
+        "split": "train",
+    },
+)
+result = run(task)
+print(json.dumps(result.json(), indent=4))
