+++ responses_api_agents/mini_swe_agent_2/data/.gitignore
@@ -1,0 +1,5 @@
+*train.jsonl
+*validation.jsonl
+*train_prepare.jsonl
+*validation_prepare.jsonl
+*example_prepare.jsonl
