+++ responses_api_agents/mini_swe_agent_2/requirements.txt
@@ -1,4 +1,6 @@
 -e nemo-gym[dev] @ ../../
-mini-swe-agent @ git+https://github.com/sdevare-nv/nv-mini-swe-agent.git@2914ef8c97b346f1ee38e3bf751a3030b0306183
+-r ../../nemo_gym/sandbox/providers/opensandbox/requirements.txt
+mini-swe-agent==2.1.0
 swegym @ git+https://github.com/sdevare-nv/nv-SWE-Bench-Package.git@31e1cb8f0241da1707d00faa633c3d6ce1a8ba3b
 docker==7.1.0
+tenacity
